rename iwidget-function
