+++ src/assessments/native-widgets/test-steps/widget-function.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import * as React from 'react';
 
-import { WidgetFunctionPropertyBag } from 'common/types/property-bag/iwidget-function';
+import { WidgetFunctionPropertyBag } from 'common/types/property-bag/widget-function-property-bag';
 import { VisualizationType } from 'common/types/visualization-type';
 import { link } from 'content/link';
 import { productName } from 'content/strings/application';
