+++ src/common/types/property-bag/widget-function-property-bag.ts
